feat: 增强网站模板管理功能，新增页面编辑器和CMS媒体库支持
- 新增模板数据按模板ID不分页查询接口
- 新增页面编辑器(page-editor)和页面抽屉(pages-drawer)组件
- 新增模板导入导出(.mtp)功能
- 扩展文件验证器支持CMS媒体库及多种内容类型
- 优化定时任务管理和国际化配置
- 新增数据库迁移脚本(v42/v44/v45)

+++ backend/src/modules/website/entity/navigation.rs
@@ -27,7 +27,7 @@ pub struct Model {
     /// 自定义url地址
     pub web_url: Option<String>,
     /// 数据 id
-    pub value: Option<u32>,
+    pub value: Option<i64>,
     /// 数据类型（custom:自定义导航, article_class:文章分类, customview:自定义页面）
     pub data_type: Option<String>,
     /// 导航类型（header:顶部导航, footer:底部导航）
